+++ example/index.js
@@ -1,25 +1,34 @@
+require('dotenv').config();
 const express = require('express');
 const path = require('path');
+// For local development replace vezgo-sdk-js vezgo ../lib/vezgo
+// const Vezgo = require('../lib/vezgo');
 const Vezgo = require('vezgo-sdk-js');
 
 const app = express();
 const port = 3001;
 
-const VEZGO_CLIENT_ID = 'YOUR_CLIENT_ID';
-const VEZGO_CLIENT_SECRET = 'YOUR_CLIENT_SECRET';
 
 const vezgo = Vezgo.init({
-  clientId: VEZGO_CLIENT_ID,
-  secret: VEZGO_CLIENT_SECRET,
+  clientId: process.env.VEZGO_CLIENT_ID,
+  secret: process.env.VEZGO_CLIENT_SECRET,
 });
 
 app.get('/assets/config.js', (req, res) => {
   res.setHeader('Content-Type', 'application/javascript');
-  res.send(`var constants = { VEZGO_CLIENT_ID: '${VEZGO_CLIENT_ID}' };`);
+  res.send(`
+    var constants = {
+      VEZGO_CLIENT_ID: '${process.env.VEZGO_CLIENT_ID}',
+      VEZGO_CONNECT_URL: '${process.env.VEZGO_CONNECT_URL || 'https://connect.vezgo.com'}',
+      VEZGO_CONNECT_TYPE: '${process.env.VEZGO_CONNECT_TYPE || 'POST'}',
+    };
+  `);
 });
 
 app.get('/assets/vezgo.js', (req, res) => {
   res.setHeader('Content-Type', 'application/javascript');
+  // For local development replace vezgo-sdk-js vezgo ../dist/vezgo.js
+  // res.sendFile(path.join(__dirname, '../dist/vezgo.js'));
   res.sendFile(path.join(__dirname, '/node_modules/vezgo-sdk-js/dist/vezgo.js'));
 });